requirements.txt
numpy==1.19.0
opencv-python==4.2.0.34
hydra-core==1.0.0rc2
tqdm==4.42.1
